+++ corehq/apps/export/views/download.py
@@ -23,7 +23,7 @@ from corehq.util.timezones.utils import get_timezone_for_user
 from corehq.toggles import MESSAGE_LOG_METADATA, PAGINATED_EXPORTS
 from corehq.apps.export.export import get_export_download, get_export_size
 from corehq.apps.export.models.new import DatePeriod, DataFile, EmailExportWhenDoneRequest
-from corehq.apps.export.views.util import ExportsPermissionsManager
+from corehq.apps.export.views.util import ExportsPermissionsManager, get_timezone
 from corehq.apps.hqwebapp.views import HQJSONResponseMixin
 from corehq.apps.hqwebapp.utils import format_angular_error, format_angular_success
 from corehq.apps.locations.models import SQLLocation
@@ -159,7 +159,7 @@ class BaseDownloadExportView(HQJSONResponseMixin, BaseProjectDataView):
     @property
     @memoized
     def timezone(self):
-        return _get_timezone(self.domain, self.request.couch_user)
+        return get_timezone(self.domain, self.request.couch_user)
 
     @property
     @memoized
