+++ corehq/tabs/tabclasses.py
@@ -497,7 +497,7 @@ class ProjectDataTab(UITab):
     @property
     @memoized
     def can_only_see_deid_exports(self):
-        from corehq.apps.export.old_views import user_can_view_deid_exports
+        from corehq.apps.export.views.util import user_can_view_deid_exports
         return (not self.can_view_form_exports
                 and user_can_view_deid_exports(self.domain, self.couch_user))
